refactor: 서버 URL / 네트워크 설정 변경 및 Shopping 패키지 내 파일 위치 이동
