+++ app/src/main/java/woowacourse/shopping/AppContainer.kt
@@ -77,7 +77,7 @@ class AppContainer(
     val retrofitService =
         Retrofit
             .Builder()
-            .baseUrl("http://techcourse-lv2-alb-974870821.ap-northeast-2.elb.amazonaws.com/")
+            .baseUrl("http://techcourse-lv2-alb-250216202.ap-northeast-2.elb.amazonaws.com/")
             .addConverterFactory(
                 Json.asConverterFactory(
                     "application/json".toMediaType(),
