+++ app/src/main/java/woowacourse/shopping/ui/shopping/component/LoadButton.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.shopping
+package woowacourse.shopping.ui.shopping.component
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Button
