+++ app/src/main/java/woowacourse/shopping/ui/shopping/component/NetworkErrorContent.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.shopping
+package woowacourse.shopping.ui.shopping.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
