+++ app/src/main/res/xml/network_security_config.xml
@@ -4,7 +4,7 @@
         <domain includeSubdomains="true">localhost</domain>
         <domain includeSubdomains="true">127.0.0.1</domain>
         <domain includeSubdomains="true">
-            techcourse-lv2-alb-974870821.ap-northeast-2.elb.amazonaws.com
+            techcourse-lv2-alb-250216202.ap-northeast-2.elb.amazonaws.com
         </domain>
     </domain-config>
 </network-security-config>
